+++ main.py
@@ -1,15 +1,10 @@
+from utils.model import toll_both
+from utils.controller import get_both_info
 
 
-toll_both : list = [
 
-    {'name' : 'SPO Olszowa', 'location' : 'Olszowa'},
-    {'name' : 'SPO Nogowczyce', 'location' : 'Nogowczyce'},
-    {'name' : 'SPO Łany', 'location' : 'Łany'},
-    {'name' : 'SPO Ostropa', 'location' : 'Ostropa'},
-]
+def main():
+    get_both_info(toll_both)
 
-def get_both_info(both_data : list) -> None:
-    for both in both_data:
-        print(f'{both['name']} : {both["location"]}')
-
-get_both_info(toll_both)
+if __name__ == '__main__':
+    main()
