--- model.py
@@ -1,2 +1,0 @@
-
-
