+++ utils/controller.py
@@ -1,0 +1,4 @@
+def get_both_info(both_data : list) -> None:
+    for both in both_data:
+        print(f'{both['name']} : {both["location"]}')
+
