+++ utils/model.py
@@ -1,0 +1,9 @@
+
+
+toll_both : list = [
+
+    {'name' : 'SPO Olszowa', 'location' : 'Olszowa'},
+    {'name' : 'SPO Nogowczyce', 'location' : 'Nogowczyce'},
+    {'name' : 'SPO Łany', 'location' : 'Łany'},
+    {'name' : 'SPO Ostropa', 'location' : 'Ostropa'},
+]
